+++ src/dto/payloads.go
@@ -3,9 +3,8 @@ package dto;
 import Entity "services.com/entity"
 type TransPayload struct {
 	CustomerId int `json:"customerId"`
-	InitialCredit int `json:"initialCredit"`
+	InitialCredit float64 `json:"initialCredit"`
 }
-
 
 type CustomerTransactionDetails struct {
 	Customer Entity.Customers `json:"customer"`
